CRUD first one

+++ src/main/java/com/example/oscarodiop/repositories/PartTypeRepository.java
@@ -2,9 +2,12 @@ package com.example.oscarodiop.repositories;
 
 import com.example.oscarodiop.models.PartType;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PartTypeRepository extends JpaRepository<PartType, Integer> {
+public interface PartTypeRepository extends JpaRepository<PartType, Long> {
+
+
 
 }
